remove swap card hook
